chore(console): refactor connectors children components
