+++ packages/console/src/pages/Connectors/components/ConnectorName/index.module.scss
